(Chore) 돔 분리
